[refactor] CustomTabView 파일로 빼기
